requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
pydantic==2.10.6
python-dotenv==1.0.1
